requirements.txt
bcrypt==3.1.7
blinker==1.4
boto3==1.14.43
botocore==1.17.43
cffi==1.14.2
click==7.1.2
cryptography==3.0
docutils==0.16
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Mail==0.9.1
Flask-PyMongo==2.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
pymongo==3.11.0
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.14.0
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
